% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_functions.R
\name{pk}
\alias{pk}
\title{Full conditional functions for all parameters}
\usage{
pk(k, kstar, lambda)
}
\arguments{
\item{k}{A scalar for the number of changes points in a cluster}

\item{kstar}{A scalar with the number maximum of change points in all clusters}

\item{lambda}{A scalar defining the parameter for the Truncate Poisson distribution
that controls the number of change points (or its initial values)}

\item{km}{A scalar for the number of changes points in a cluster}

\item{mk}{A matrix with all possible values to distribute between change points}

\item{m0}{A scalar for the number of positions available to define change-points positions}

\item{Yn}{A vector or matrix with data sequences for a cluster}

\item{Tln}{A vector with the change-point positions for a cluster}

\item{alphan}{A vector with the constant level values for each interval between change points for a cluster}

\item{sigma2n}{A vector with the variance of the data sequences in a cluster}

\item{cellsn}{A vector with the indices of the data sequences in a cluster}

\item{Cr}{A scalar with the number of data sequences in a cluster}

\item{clusteri}{A scalar with the index of a cluster}

\item{as}{The hyperparameter value for the shape parameter in the inverse-gamma prior for the variance
component}

\item{bs}{The hyperparameter value for the scale parameter in the inverse-gamma prior for the variance
component}

\item{M}{A scalar representing the number of points available for each data sequence}

\item{w}{A scalar representing the minimum number of points in each interval between two change points}

\item{N}{A scalar representing the number of data sequences}

\item{al}{The hyperparameter value for the shape parameter in the gamma prior for lambda}

\item{bl}{The hyperparameter value for the scale parameter in the gamma prior for lambda}

\item{a}{The hyperparameter value for the shape parameter in the gamma prior for alpha0}

\item{b}{The hyperparameter value for the scale parameter in the gamma prior for alpha0}

\item{alpha0}{A scalar defining the parameter for the Dirichlet process prior
that controls the number of clusters (or its initial values)}

\item{K}{A vector containing the number of change points for each cluster (or its initial values)}

\item{Tl}{A list containing a vector for each cluster determining the change-point positions in each cluster
(or its initial values)}

\item{cluster}{A vector containing the cluster assignments for the data sequences (or its initial values)}

\item{alpha}{A list containing a vector for each cluster determining the constant level values
for each interval between change points in each cluster (or its initial values)}

\item{sigma2}{A vector with the variances of the data sequences (or its initial values)}
}
\description{
Full conditional functions for all parameters
}
