% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gibbs_true.R
\name{run_gibbs_true}
\alias{run_gibbs_true}
\title{Simulates a dataset using `simdata` function and runs the Gibbs sampler algorithm
   for simulated scenarios using true values as the initial values}
\usage{
run_gibbs_true(
  M,
  N,
  w,
  d,
  as = 2,
  bs = 100,
  al = 2,
  bl = 1000,
  a = 2,
  b = 1000,
  alpha0 = 1/100,
  lambda = 2,
  maxIter = 10000,
  par.values,
  emp,
  seed
)
}
\arguments{
\item{M}{A scalar representing the number of points available for each observation}

\item{N}{A scalar representing the number of observations}

\item{w}{A scalar representing the minimum number of points in each interval between two change points}

\item{d}{A scalar representing the number of clusters.}

\item{as}{The hyperparameter value for the shape parameter in the inverse-gamma prior for the variance component}

\item{bs}{The hyperparameter value for the scale parameter in the inverse-gamma prior for the variance component}

\item{al}{The hyperparameter value for the shape parameter in the gamma prior for lambda}

\item{bl}{The hyperparameter value for the scale parameter in the gamma prior for lambda}

\item{a}{The hyperparameter value for the shape parameter in the gamma prior for alpha0}

\item{b}{The hyperparameter value for the scale parameter in the gamma prior for alpha0}

\item{alpha0}{A scalar defining the parameter for the Dirichlet process prior
that controls the number of clusters (or its initial values)}

\item{lambda}{A scalar defining the parameter for the Truncate Poisson distribution
that controls the number of change points (or its initial values)}

\item{maxIter}{A scalar for the number of iteration to run in the Gibbs sampler}

\item{par.values}{A list with lists with parameters for each cluster.
The first argument in each list is the number of change points, then the positions for the change points,
where $T_1 = 1, T_{last} = M + 1$, and for each interval between change points
you need to specify a value for the constant level.}

\item{emp}{A boolean value (TRUE/FALSE) for the usage of empirical variance. Default is set to `FALSE`}

\item{seed}{A scalar representing the seed}
}
\value{
A list with the output from the `simdata` function and the estimates for each iteration
    of the Gibbs sampler for each parameter
}
\description{
Simulates a dataset using `simdata` function and runs the Gibbs sampler algorithm
   for simulated scenarios using true values as the initial values
}
