% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_random.R
\name{sim_random}
\alias{sim_random}
\title{Run and saves the Gibbs sampler algorithm results for simulated scenarios or for real datasets
   when using initial_values.}
\usage{
sim_random(
  S,
  seed,
  scenario,
  casenumber,
  M,
  N,
  w,
  d,
  as,
  bs,
  al,
  bl,
  a,
  b,
  alpha0,
  lambda,
  maxIter,
  startpoint_sampling,
  par.values,
  sigma2_equal = FALSE,
  emp = FALSE,
  initial_val
)
}
\arguments{
\item{S}{A scalar representing the number of simulations to run.}

\item{seed}{A scalar representing the seed}

\item{scenario}{A string with the name of the folder you want to create to save the results}

\item{casenumber}{A scalar with the id for the simulation scenario you want to run.
If only one simulation will be performed set this argument to 1.}

\item{M}{A scalar representing the number of points available for each observation}

\item{N}{A scalar representing the number of observations}

\item{w}{A scalar representing the minimum number of points in each interval between two change points}

\item{d}{A scalar representing the number of clusters.}

\item{as}{The hyperparameter value for the shape parameter in the inverse-gamma prior for the variance component}

\item{bs}{The hyperparameter value for the scale parameter in the inverse-gamma prior for the variance component}

\item{al}{The hyperparameter value for the shape parameter in the gamma prior for lambda}

\item{bl}{The hyperparameter value for the scale parameter in the gamma prior for lambda}

\item{a}{The hyperparameter value for the shape parameter in the gamma prior for alpha0}

\item{b}{The hyperparameter value for the scale parameter in the gamma prior for alpha0}

\item{alpha0}{A scalar defining the parameter for the Dirichlet process prior
that controls the number of clusters}

\item{lambda}{A scalar defining the parameter for the Truncate Poisson distribution
that controls the number of change points}

\item{maxIter}{A scalar for the number of iteration to run in the Gibbs sampler}

\item{startpoint_sampling}{A scalar defining the step value to use when selecting a systematic sample
after the burn-in.}

\item{par.values}{A list with lists with parameters for each cluster.
The first argument in each list is the number of change points, then the positions for the change points,
where $T_1 = 1, T_{last} = M + 1$, and for each interval between change points
you need to specify a value for the constant level.}

\item{sigma2_equal}{A boolean value (TRUE/FALSE) to determine if the observations should have the same variance.
Default is set to `FALSE`}

\item{emp}{A boolean value (TRUE/FALSE) for the usage of empirical variance. Default is set to `FALSE`}

\item{initial_val}{A list with the same structure as the `par.values` argument specifying the initial values
for each parameter to estimate}
}
\value{
It saves the posterior estimates in the folder provided and returns the processing time.
}
\description{
Run and saves the Gibbs sampler algorithm results for simulated scenarios or for real datasets
   when using initial_values.
}
