% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{simdata}
\alias{simdata}
\title{Simulate change-point data}
\usage{
simdata(par.values, d, M, w, N, as, bs, emp, seed)
}
\arguments{
\item{par.values}{A list with a list with parameter for each cluster in the data vector with, at most, one element.}

\item{d}{A scalar representing the number of clusters.}

\item{M}{A scalar representing the number of points available for each observation}

\item{w}{A scalar representing the minimum number of points in each interval between two change points}

\item{N}{A scalar representing the number of observations}

\item{as}{The hyperparameter value for the shape parameter in the inverse-gamma prior for the variance component}

\item{bs}{The hyperparameter value for the scale parameter in the inverse-gamma prior for the variance component}

\item{emp}{A boolean value (TRUE/FALSE) for the usage of empirical variance}

\item{seed}{A scalar representing the seed}
}
\value{
A list composed of a matrix with the data, a vector with the cluster assignments for each observation,
   and a vector with the variance for each observation.
}
\description{
Simulate change-point data
}
\examples{
set.seed(1238)
sigma2 <- round(rinvgamma(10, 21, 0.05*20),4)
param1 <- list(list(2, c(1, 19, 34, 51), c(5, 20, 10)), list(2, c(1, 15, 32, 51), c(17, 10, 2)), sigma2)
simdata(par.values = param1, d = 2, M = 50, w = 10, N = 10, as = 2, bs = 1000, emp = FALSE, seed = 1252)
}
